Move project explorer tree expander to dedicate class
